refactor: Update import paths for MultipleSelector and related components to use new registry structure
