+++ app/registry/multiple-selector/usage/multiple-selector-controlled.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React from 'react';
-import MultipleSelector, { Option } from '@/components/ui/multiple-selector';
+import MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';
 
 const OPTIONS: Option[] = [
   { label: 'nextjs', value: 'Nextjs' },
