+++ app/registry/multiple-selector/usage/multiple-selector-max-text-length.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MultipleSelector, { Option } from '@/components/ui/multiple-selector';
+import MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';
 
 const OPTIONS: Option[] = [
   { label: 'nextjs', value: 'nextjs' },
