+++ public/r/multiple-selector-commandprops.json
@@ -5,13 +5,13 @@
   "title": "Multiple Selector Command Props",
   "description": "component for the Multiple Selector Command Props",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json",
+    "http://localhost:3000/r/multiple-selector-dependencies.json",
     "http://localhost:3000/r/inline-code.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-commandprops.tsx",
-      "content": "'use client';\r\nimport React, { useState } from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\nimport { InlineCode } from '@/components/ui/inline-code';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorCommandProps = () => {\r\n  const [isFocus, setIsFocus] = useState(false);\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <p>\r\n        Is focus? <InlineCode>{String(isFocus)}</InlineCode>\r\n      </p>\r\n      <p>\r\n        You can open console and search <InlineCode>command description</InlineCode> and{' '}\r\n        <InlineCode>input description</InlineCode> to check the customization of{' '}\r\n        <InlineCode>aria-description</InlineCode>.\r\n      </p>\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n        commandProps={{ 'aria-description': 'command description' }}\r\n        inputProps={{\r\n          onBlur: () => {\r\n            setIsFocus(false);\r\n          },\r\n          onFocus: () => {\r\n            setIsFocus(true);\r\n          },\r\n          'aria-description': 'input description',\r\n        }}\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorCommandProps;\r\n",
+      "content": "'use client';\r\nimport React, { useState } from 'react';\r\n\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\n\r\nimport { InlineCode } from '@/app/registry/spectrumui/inline-code';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorCommandProps = () => {\r\n  const [isFocus, setIsFocus] = useState(false);\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <p>\r\n        Is focus? <InlineCode>{String(isFocus)}</InlineCode>\r\n      </p>\r\n      <p>\r\n        You can open console and search <InlineCode>command description</InlineCode> and{' '}\r\n        <InlineCode>input description</InlineCode> to check the customization of{' '}\r\n        <InlineCode>aria-description</InlineCode>.\r\n      </p>\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n        commandProps={{ 'aria-description': 'command description' }}\r\n        inputProps={{\r\n          onBlur: () => {\r\n            setIsFocus(false);\r\n          },\r\n          onFocus: () => {\r\n            setIsFocus(true);\r\n          },\r\n          'aria-description': 'input description',\r\n        }}\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorCommandProps;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-commandprops.tsx"
     }
