+++ public/r/multiple-selector-controlled.json
@@ -5,12 +5,12 @@
   "title": "Multiple Selector Controlled",
   "description": "component for the Multiple Selector Controlled",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json"
+    "http://localhost:3000/r/multiple-selector-dependencies.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-controlled.tsx",
-      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember' },\r\n  { label: 'Gatsby', value: 'gatsby' },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorControlled = () => {\r\n  const [value, setValue] = React.useState<Option[]>([]);\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <p className=\"text-primary\">Your selection: {value.map((val) => val.label).join(', ')}</p>\r\n      <MultipleSelector\r\n        value={value}\r\n        onChange={setValue}\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorControlled;\r\n",
+      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember' },\r\n  { label: 'Gatsby', value: 'gatsby' },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorControlled = () => {\r\n  const [value, setValue] = React.useState<Option[]>([]);\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <p className=\"text-primary\">Your selection: {value.map((val) => val.label).join(', ')}</p>\r\n      <MultipleSelector\r\n        value={value}\r\n        onChange={setValue}\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorControlled;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-controlled.tsx"
     }
