+++ public/r/multiple-selector-fixed-option.json
@@ -5,12 +5,12 @@
   "title": "Multiple Selector Fixed Option",
   "description": "component for the Multiple Selector Fixed Option",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json"
+    "http://localhost:3000/r/multiple-selector-dependencies.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-fixed-option.tsx",
-      "content": "'use client';\r\nimport React, { useState } from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorWithFixedOption = () => {\r\n  const [value, setValue] = useState<Option[]>([\r\n    { label: 'nextjs', value: 'nextjs', fixed: true },\r\n    { label: 'React', value: 'react', fixed: true },\r\n  ]);\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <MultipleSelector\r\n        value={value}\r\n        onChange={setValue}\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithFixedOption;\r\n",
+      "content": "'use client';\r\nimport React, { useState } from 'react';\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorWithFixedOption = () => {\r\n  const [value, setValue] = useState<Option[]>([\r\n    { label: 'nextjs', value: 'nextjs', fixed: true },\r\n    { label: 'React', value: 'react', fixed: true },\r\n  ]);\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <MultipleSelector\r\n        value={value}\r\n        onChange={setValue}\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithFixedOption;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-fixed-option.tsx"
     }
