+++ public/r/multiple-selector-manually-controlled.json
@@ -5,13 +5,13 @@
   "title": "Multiple Selector Manually Controlled",
   "description": "component for the Multiple Selector Manually Controlled",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json",
+    "http://localhost:3000/r/multiple-selector-dependencies.json",
     "http://localhost:3000/r/loading_button_dependencies.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-manually-controlled-options.tsx",
-      "content": "'use client';\r\nimport React, { useState } from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\nimport { LoadingButton } from '@/components/ui/loading-button';\r\n\r\nconst REACT_OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'shadcn-ui', value: 'shadcn-ui' },\r\n  { label: 'MUI', value: 'mui' },\r\n  { label: 'Ant Design', value: 'antDesign' },\r\n];\r\n\r\nconst VUE_OPTIONS: Option[] = [\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Vuetify', value: 'vuetify' },\r\n  { label: 'Quasar', value: 'quasar' },\r\n];\r\n\r\nconst MultipleSelectorManuallyControlledOptions = () => {\r\n  const [options, setOptions] = useState<Option[]>(REACT_OPTIONS);\r\n  const [placeholder, setPlaceholder] = useState('Which framework do you want to use for react?');\r\n  const [toggle, setToggle] = useState('react');\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <LoadingButton\r\n        className=\"mb-3\"\r\n        onClick={() => {\r\n          if (toggle === 'react') {\r\n            setOptions(VUE_OPTIONS);\r\n            setPlaceholder('Which framework do you want to use for vue?');\r\n            setToggle('vue');\r\n          } else {\r\n            setOptions(REACT_OPTIONS);\r\n            setPlaceholder('Which framework do you want to use for react?');\r\n            setToggle('react');\r\n          }\r\n        }}\r\n      >\r\n        change options to {toggle === 'react' ? 'vue' : 'react'}\r\n      </LoadingButton>\r\n      <MultipleSelector\r\n        options={options}\r\n        placeholder={placeholder}\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorManuallyControlledOptions;\r\n",
+      "content": "'use client';\r\nimport React, { useState } from 'react';\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\n\r\nimport { LoadingButton } from '@/app/registry/spectrumui/loading-button-dependencies';\r\n\r\nconst REACT_OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'shadcn-ui', value: 'shadcn-ui' },\r\n  { label: 'MUI', value: 'mui' },\r\n  { label: 'Ant Design', value: 'antDesign' },\r\n];\r\n\r\nconst VUE_OPTIONS: Option[] = [\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Vuetify', value: 'vuetify' },\r\n  { label: 'Quasar', value: 'quasar' },\r\n];\r\n\r\nconst MultipleSelectorManuallyControlledOptions = () => {\r\n  const [options, setOptions] = useState<Option[]>(REACT_OPTIONS);\r\n  const [placeholder, setPlaceholder] = useState('Which framework do you want to use for react?');\r\n  const [toggle, setToggle] = useState('react');\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <LoadingButton\r\n        className=\"mb-3\"\r\n        onClick={() => {\r\n          if (toggle === 'react') {\r\n            setOptions(VUE_OPTIONS);\r\n            setPlaceholder('Which framework do you want to use for vue?');\r\n            setToggle('vue');\r\n          } else {\r\n            setOptions(REACT_OPTIONS);\r\n            setPlaceholder('Which framework do you want to use for react?');\r\n            setToggle('react');\r\n          }\r\n        }}\r\n      >\r\n        change options to {toggle === 'react' ? 'vue' : 'react'}\r\n      </LoadingButton>\r\n      <MultipleSelector\r\n        options={options}\r\n        placeholder={placeholder}\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorManuallyControlledOptions;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-manually-controlled.tsx"
     }
