+++ public/r/multiple-selector-max-selected.json
@@ -5,13 +5,13 @@
   "title": "Multiple Selector Max Selected",
   "description": "component for the Multiple Selector Max Selected",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json",
+    "http://localhost:3000/r/multiple-selector-dependencies.json",
     "http://localhost:3000/r/use-toast.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-max-selected.tsx",
-      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\nimport { toast } from '@/components/ui/use-toast';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember' },\r\n  { label: 'Gatsby', value: 'gatsby' },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorWithMaxSelected = () => {\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <MultipleSelector\r\n        maxSelected={3}\r\n        onMaxSelected={(maxLimit) => {\r\n          toast({\r\n            title: `You have reached max selected: ${maxLimit}`,\r\n          });\r\n        }}\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithMaxSelected;\r\n",
+      "content": "'use client';\r\nimport React from 'react';\r\n\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\nimport { toast } from '@/app/registry/spectrumui/use-toast';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember' },\r\n  { label: 'Gatsby', value: 'gatsby' },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorWithMaxSelected = () => {\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <MultipleSelector\r\n        maxSelected={3}\r\n        onMaxSelected={(maxLimit) => {\r\n          toast({\r\n            title: `You have reached max selected: ${maxLimit}`,\r\n          });\r\n        }}\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithMaxSelected;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-max-selected.tsx"
     }
