+++ public/r/multiple-selector-no-placeholder-when-selected.json
@@ -5,12 +5,12 @@
   "title": "Multiple Selector No Placeholder When Selected",
   "description": "component for the Multiple Selector No Placeholder When Selected",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json"
+    "http://localhost:3000/r/multiple-selector-dependencies.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-no-placeholder-when-selected.tsx",
-      "content": "import React from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorNoPlaceholderWhenSelected = () => {\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        hidePlaceholderWhenSelected\r\n        placeholder=\"The placeholder will be hidden when there are options selected.\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorNoPlaceholderWhenSelected;\r\n",
+      "content": "import React from 'react';\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorNoPlaceholderWhenSelected = () => {\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        hidePlaceholderWhenSelected\r\n        placeholder=\"The placeholder will be hidden when there are options selected.\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorNoPlaceholderWhenSelected;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-no-placeholder-when-selected.tsx"
     }
