+++ public/r/multiple-selector-ref.json
@@ -5,13 +5,14 @@
   "title": "Multiple Selector Ref",
   "description": "component for the Multiple Selector Ref",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json",
-    "http://localhost:3000/r/use-toast.json"
+    "http://localhost:3000/r/multiple-selector-dependencies.json",
+    "http://localhost:3000/r/use-toast.json",
+    "http://localhost:3000/r/loading_button_dependencies.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-ref.tsx",
-      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { type MultipleSelectorRef, Option } from '@/components/ui/multiple-selector';\r\nimport { LoadingButton } from '@/components/ui/loading-button';\r\nimport { toast } from '@/components/ui/use-toast';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorRef = () => {\r\n  const ref = React.useRef<MultipleSelectorRef>(null);\r\n\r\n  /**\r\n   * An example for focus on input when component mounted.\r\n   * I commented it out because it will affect the demo page.\r\n   **/\r\n  // useEffect(() => {\r\n  //   if (ref.current) {\r\n  //     ref.current.input.focus();\r\n  //   }\r\n  // }, []);\r\n\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        ref={ref}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n      <div className=\"flex gap-3\">\r\n        <LoadingButton\r\n          onClick={() => {\r\n            toast({\r\n              title: 'Your ref data',\r\n              description: (\r\n                <pre className=\"mt-2 w-[340px] rounded-md bg-slate-950 p-4\">\r\n                  <code className=\"text-white\">\r\n                    {JSON.stringify(ref.current?.selectedValue, null, 2)}\r\n                  </code>\r\n                </pre>\r\n              ),\r\n            });\r\n          }}\r\n        >\r\n          get value from ref\r\n        </LoadingButton>\r\n        <LoadingButton\r\n          onClick={() => {\r\n            ref.current?.input.focus();\r\n          }}\r\n        >\r\n          focus on multiple selector\r\n        </LoadingButton>\r\n      </div>\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorRef;\r\n",
+      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { type MultipleSelectorRef,Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\nimport { LoadingButton } from '@/app/registry/spectrumui/loading-button-dependencies';\r\nimport { toast } from '@/app/registry/spectrumui/use-toast';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst MultipleSelectorRef = () => {\r\n  const ref = React.useRef<MultipleSelectorRef>(null);\r\n\r\n  /**\r\n   * An example for focus on input when component mounted.\r\n   * I commented it out because it will affect the demo page.\r\n   **/\r\n  // useEffect(() => {\r\n  //   if (ref.current) {\r\n  //     ref.current.input.focus();\r\n  //   }\r\n  // }, []);\r\n\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        ref={ref}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n      <div className=\"flex gap-3\">\r\n        <LoadingButton\r\n          onClick={() => {\r\n            toast({\r\n              title: 'Your ref data',\r\n              description: (\r\n                <pre className=\"mt-2 w-[340px] rounded-md bg-slate-950 p-4\">\r\n                  <code className=\"text-white\">\r\n                    {JSON.stringify(ref.current?.selectedValue, null, 2)}\r\n                  </code>\r\n                </pre>\r\n              ),\r\n            });\r\n          }}\r\n        >\r\n          get value from ref\r\n        </LoadingButton>\r\n        <LoadingButton\r\n          onClick={() => {\r\n            ref.current?.input.focus();\r\n          }}\r\n        >\r\n          focus on multiple selector\r\n        </LoadingButton>\r\n      </div>\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorRef;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-ref.tsx"
     }
