+++ public/r/multiple-selector-with-async-search-and-creatable-and-group.json
@@ -5,12 +5,13 @@
   "title": "Multiple Selector With Async Search And Creatable And Group",
   "description": "component for the Multiple Selector With Async Search And Creatable And Group",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json"
+    "http://localhost:3000/r/multiple-selector-dependencies.json",
+    "http://localhost:3000/r/inline-code.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-with-async-search-and-creatable-and-group.tsx",
-      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\nimport { InlineCode } from '@/components/ui/inline-code';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs', group: 'React' },\r\n  { label: 'React', value: 'react', group: 'React' },\r\n  { label: 'Remix', value: 'remix', group: 'React' },\r\n  { label: 'shadcn-ui', value: 'shadcn-ui', group: 'React' },\r\n  { label: 'mui', value: 'mui', group: 'React' },\r\n  { label: 'Vite', value: 'vite', group: 'Vue' },\r\n  { label: 'Nuxt', value: 'nuxt', group: 'Vue' },\r\n  { label: 'Vue', value: 'vue', group: 'Vue' },\r\n  { label: 'Quasar', value: 'quasar', group: 'Vue' },\r\n  { label: 'Angular', value: 'angular', group: 'Angular' },\r\n  { label: 'Material UI', value: 'material-ui', group: 'Angular' },\r\n  { label: 'Ng-zorro', value: 'ng-zorro', group: 'Angular' },\r\n];\r\n\r\nconst mockSearch = async (value: string): Promise<Option[]> => {\r\n  return new Promise((resolve) => {\r\n    setTimeout(() => {\r\n      const res = OPTIONS.filter((option) => option.value.includes(value));\r\n      resolve(res);\r\n    }, 1000);\r\n  });\r\n};\r\n\r\nconst MultipleSelectorWithAsyncSearchAndCreatableAndGroup = () => {\r\n  const [isTriggered, setIsTriggered] = React.useState(false);\r\n\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <p>\r\n        Is request been triggered? <InlineCode>{String(isTriggered)}</InlineCode>\r\n      </p>\r\n      <MultipleSelector\r\n        onSearch={async (value) => {\r\n          setIsTriggered(true);\r\n          const res = await mockSearch(value);\r\n          setIsTriggered(false);\r\n          return res;\r\n        }}\r\n        defaultOptions={[]}\r\n        creatable\r\n        groupBy=\"group\"\r\n        placeholder=\"trying to search 'a' to get more options...\"\r\n        loadingIndicator={\r\n          <p className=\"py-2 text-center text-lg leading-10 text-muted-foreground\">loading...</p>\r\n        }\r\n        emptyIndicator={\r\n          <p className=\"w-full text-center text-lg leading-10 text-muted-foreground\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithAsyncSearchAndCreatableAndGroup;\r\n",
+      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\nimport { InlineCode } from '@/app/registry/spectrumui/inline-code';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs', group: 'React' },\r\n  { label: 'React', value: 'react', group: 'React' },\r\n  { label: 'Remix', value: 'remix', group: 'React' },\r\n  { label: 'shadcn-ui', value: 'shadcn-ui', group: 'React' },\r\n  { label: 'mui', value: 'mui', group: 'React' },\r\n  { label: 'Vite', value: 'vite', group: 'Vue' },\r\n  { label: 'Nuxt', value: 'nuxt', group: 'Vue' },\r\n  { label: 'Vue', value: 'vue', group: 'Vue' },\r\n  { label: 'Quasar', value: 'quasar', group: 'Vue' },\r\n  { label: 'Angular', value: 'angular', group: 'Angular' },\r\n  { label: 'Material UI', value: 'material-ui', group: 'Angular' },\r\n  { label: 'Ng-zorro', value: 'ng-zorro', group: 'Angular' },\r\n];\r\n\r\nconst mockSearch = async (value: string): Promise<Option[]> => {\r\n  return new Promise((resolve) => {\r\n    setTimeout(() => {\r\n      const res = OPTIONS.filter((option) => option.value.includes(value));\r\n      resolve(res);\r\n    }, 1000);\r\n  });\r\n};\r\n\r\nconst MultipleSelectorWithAsyncSearchAndCreatableAndGroup = () => {\r\n  const [isTriggered, setIsTriggered] = React.useState(false);\r\n\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <p>\r\n        Is request been triggered? <InlineCode>{String(isTriggered)}</InlineCode>\r\n      </p>\r\n      <MultipleSelector\r\n        onSearch={async (value) => {\r\n          setIsTriggered(true);\r\n          const res = await mockSearch(value);\r\n          setIsTriggered(false);\r\n          return res;\r\n        }}\r\n        defaultOptions={[]}\r\n        creatable\r\n        groupBy=\"group\"\r\n        placeholder=\"trying to search 'a' to get more options...\"\r\n        loadingIndicator={\r\n          <p className=\"py-2 text-center text-lg leading-10 text-muted-foreground\">loading...</p>\r\n        }\r\n        emptyIndicator={\r\n          <p className=\"w-full text-center text-lg leading-10 text-muted-foreground\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithAsyncSearchAndCreatableAndGroup;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-with-async-search-and-creatable-and-group.tsx"
     }
