+++ public/r/multiple-selector-with-async-search-and-creatable.json
@@ -5,12 +5,13 @@
   "title": "Multiple Selector With Async Search And Creatable",
   "description": "component for the Multiple Selector With Async Search And Creatable",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json"
+    "http://localhost:3000/r/multiple-selector-dependencies.json",
+    "http://localhost:3000/r/inline-code.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-with-async-search-and-creatable.tsx",
-      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\nimport { InlineCode } from '@/components/ui/inline-code';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember' },\r\n  { label: 'Gatsby', value: 'gatsby' },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst mockSearch = async (value: string): Promise<Option[]> => {\r\n  return new Promise((resolve) => {\r\n    setTimeout(() => {\r\n      const res = OPTIONS.filter((option) => option.value.includes(value));\r\n      resolve(res);\r\n    }, 1000);\r\n  });\r\n};\r\n\r\nconst MultipleSelectorWithAsyncSearchAndCreatable = () => {\r\n  const [isTriggered, setIsTriggered] = React.useState(false);\r\n\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <p>\r\n        Is request been triggered? <InlineCode>{String(isTriggered)}</InlineCode>\r\n      </p>\r\n      <MultipleSelector\r\n        onSearch={async (value) => {\r\n          setIsTriggered(true);\r\n          const res = await mockSearch(value);\r\n          setIsTriggered(false);\r\n          return res;\r\n        }}\r\n        defaultOptions={[]}\r\n        creatable\r\n        placeholder=\"trying to search 'a' to get more options...\"\r\n        loadingIndicator={\r\n          <p className=\"py-2 text-center text-lg leading-10 text-muted-foreground\">loading...</p>\r\n        }\r\n        emptyIndicator={\r\n          <p className=\"w-full text-center text-lg leading-10 text-muted-foreground\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithAsyncSearchAndCreatable;\r\n",
+      "content": "'use client';\r\nimport React from 'react';\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\nimport { InlineCode } from '@/app/registry/spectrumui/inline-code';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Vite', value: 'vite' },\r\n  { label: 'Nuxt', value: 'nuxt' },\r\n  { label: 'Vue', value: 'vue' },\r\n  { label: 'Svelte', value: 'svelte' },\r\n  { label: 'Angular', value: 'angular' },\r\n  { label: 'Ember', value: 'ember' },\r\n  { label: 'Gatsby', value: 'gatsby' },\r\n  { label: 'Astro', value: 'astro' },\r\n];\r\n\r\nconst mockSearch = async (value: string): Promise<Option[]> => {\r\n  return new Promise((resolve) => {\r\n    setTimeout(() => {\r\n      const res = OPTIONS.filter((option) => option.value.includes(value));\r\n      resolve(res);\r\n    }, 1000);\r\n  });\r\n};\r\n\r\nconst MultipleSelectorWithAsyncSearchAndCreatable = () => {\r\n  const [isTriggered, setIsTriggered] = React.useState(false);\r\n\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <p>\r\n        Is request been triggered? <InlineCode>{String(isTriggered)}</InlineCode>\r\n      </p>\r\n      <MultipleSelector\r\n        onSearch={async (value) => {\r\n          setIsTriggered(true);\r\n          const res = await mockSearch(value);\r\n          setIsTriggered(false);\r\n          return res;\r\n        }}\r\n        defaultOptions={[]}\r\n        creatable\r\n        placeholder=\"trying to search 'a' to get more options...\"\r\n        loadingIndicator={\r\n          <p className=\"py-2 text-center text-lg leading-10 text-muted-foreground\">loading...</p>\r\n        }\r\n        emptyIndicator={\r\n          <p className=\"w-full text-center text-lg leading-10 text-muted-foreground\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithAsyncSearchAndCreatable;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-with-async-search-and-creatable.tsx"
     }
