+++ public/r/multiple-selector-with-disable-options.json
@@ -5,12 +5,12 @@
   "title": "Multiple Selector With Disable Options",
   "description": "component for the Multiple Selector With Disable Options",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json"
+    "http://localhost:3000/r/multiple-selector-dependencies.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-with-disable-option.tsx",
-      "content": "import React from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'Vite', value: 'vite', disable: true },\r\n  { label: 'Nuxt', value: 'nuxt', disable: true },\r\n  { label: 'Vue', value: 'vue, disable: true', disable: true },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Svelte', value: 'svelte', disable: true },\r\n  { label: 'Angular', value: 'angular', disable: true },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro', disable: true },\r\n];\r\n\r\nconst MultipleSelectorWithDisabledOption = () => {\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithDisabledOption;\r\n",
+      "content": "import React from 'react';\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\n\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'Nextjs' },\r\n  { label: 'Vite', value: 'vite', disable: true },\r\n  { label: 'Nuxt', value: 'nuxt', disable: true },\r\n  { label: 'Vue', value: 'vue, disable: true', disable: true },\r\n  { label: 'Remix', value: 'remix' },\r\n  { label: 'Svelte', value: 'svelte', disable: true },\r\n  { label: 'Angular', value: 'angular', disable: true },\r\n  { label: 'Ember', value: 'ember', disable: true },\r\n  { label: 'React', value: 'react' },\r\n  { label: 'Gatsby', value: 'gatsby', disable: true },\r\n  { label: 'Astro', value: 'astro', disable: true },\r\n];\r\n\r\nconst MultipleSelectorWithDisabledOption = () => {\r\n  return (\r\n    <div className=\"flex w-full flex-col gap-5 px-10\">\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithDisabledOption;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-with-disable-option.tsx"
     }
