+++ public/r/multiple-selector-with-group.json
@@ -5,12 +5,12 @@
   "title": "Multiple Selector With Group",
   "description": "component for the Multiple Selector With Group",
   "registryDependencies": [
-    "http://localhost:3000/r/multiple_selector_dependencies.json"
+    "http://localhost:3000/r/multiple-selector-dependencies.json"
   ],
   "files": [
     {
       "path": "app/registry/multiple-selector/usage/multiple-selector-with-group.tsx",
-      "content": "import React from 'react';\r\nimport MultipleSelector, { Option } from '@/components/ui/multiple-selector';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs', group: 'React' },\r\n  { label: 'React', value: 'react', group: 'React' },\r\n  { label: 'Remix', value: 'remix', group: 'React' },\r\n  { label: 'shadcn-ui', value: 'shadcn-ui', group: 'React' },\r\n  { label: 'mui', value: 'mui', group: 'React' },\r\n  { label: 'Vite', value: 'vite', group: 'Vue' },\r\n  { label: 'Nuxt', value: 'nuxt', group: 'Vue' },\r\n  { label: 'Vue', value: 'vue', group: 'Vue' },\r\n  { label: 'Quasar', value: 'quasar', group: 'Vue' },\r\n  { label: 'Angular', value: 'angular', group: 'Angular' },\r\n  { label: 'Material UI', value: 'material-ui', group: 'Angular' },\r\n  { label: 'Ng-zorro', value: 'ng-zorro', group: 'Angular' },\r\n];\r\n\r\nconst MultipleSelectorWithGroup = () => {\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n        groupBy=\"group\"\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithGroup;\r\n",
+      "content": "import React from 'react';\r\nimport MultipleSelector, { Option } from '@/app/registry/spectrumui/multiple-selector-dependencies';\r\n\r\nconst OPTIONS: Option[] = [\r\n  { label: 'nextjs', value: 'nextjs', group: 'React' },\r\n  { label: 'React', value: 'react', group: 'React' },\r\n  { label: 'Remix', value: 'remix', group: 'React' },\r\n  { label: 'shadcn-ui', value: 'shadcn-ui', group: 'React' },\r\n  { label: 'mui', value: 'mui', group: 'React' },\r\n  { label: 'Vite', value: 'vite', group: 'Vue' },\r\n  { label: 'Nuxt', value: 'nuxt', group: 'Vue' },\r\n  { label: 'Vue', value: 'vue', group: 'Vue' },\r\n  { label: 'Quasar', value: 'quasar', group: 'Vue' },\r\n  { label: 'Angular', value: 'angular', group: 'Angular' },\r\n  { label: 'Material UI', value: 'material-ui', group: 'Angular' },\r\n  { label: 'Ng-zorro', value: 'ng-zorro', group: 'Angular' },\r\n];\r\n\r\nconst MultipleSelectorWithGroup = () => {\r\n  return (\r\n    <div className=\"w-full px-10\">\r\n      <MultipleSelector\r\n        defaultOptions={OPTIONS}\r\n        placeholder=\"Select frameworks you like...\"\r\n        emptyIndicator={\r\n          <p className=\"text-center text-lg leading-10 text-gray-600 dark:text-gray-400\">\r\n            no results found.\r\n          </p>\r\n        }\r\n        groupBy=\"group\"\r\n      />\r\n    </div>\r\n  );\r\n};\r\n\r\nexport default MultipleSelectorWithGroup;\r\n",
       "type": "registry:ui",
       "target": "components/spectrumui/multiple-selector-with-group.tsx"
     }
